fine tune datatype

+++ NeoPixelMatrix_rainbow_ATtiny85/light_ws2812_1bitPixel.h
@@ -94,8 +94,8 @@
 #define w_nop8 w_nop4 w_nop4
 #define w_nop16 w_nop8 w_nop8
 
-typedef uint8_t (*getPixelColor)(uint8_t x, uint8_t y, uint8_t bgrIdx);
-void inline ws2812_set_leds_func_ptr_mask(uint8_t w, uint8_t h, getPixelColor getPixelColorPtr, uint8_t maskhi)
+typedef uint8_t (*getPixelColor)(uint16_t x, uint8_t y, uint8_t bgrIdx);
+void inline ws2812_set_leds_func_ptr_mask(uint16_t w, uint8_t h, getPixelColor getPixelColorPtr, uint8_t maskhi)
 {
   uint8_t curbyte, ctr, masklo;
   uint8_t sreg_prev;
@@ -110,7 +110,7 @@ void inline ws2812_set_leds_func_ptr_mask(uint8_t w, uint8_t h, getPixelColor ge
 
   for (uint16_t x = 0; x < w; ++x)
   {
-    for (uint16_t y = 0; y < h; ++y)
+    for (uint8_t y = 0; y < h; ++y)
     {
       uint8_t brg_idx = 3;
       while (brg_idx--)
@@ -180,7 +180,7 @@ void inline ws2812_set_leds_func_ptr_mask(uint8_t w, uint8_t h, getPixelColor ge
   SREG = sreg_prev;
 }
 
-void ws2812_set_leds_func_ptr(uint8_t width, uint8_t height, getPixelColor getPixelColorPtr)
+void ws2812_set_leds_func_ptr(uint16_t width, uint8_t height, getPixelColor getPixelColorPtr)
 {
   ws2812_set_leds_func_ptr_mask(width, height, getPixelColorPtr, _BV(ws2812_pin));
 }
